separated interactive widgets to separate files
