+++ pytermgui/widgets/interactive/__init__.py
@@ -1,0 +1,7 @@
+"""The module housing the various interactive Widget types."""
+
+from .button import Button
+from .checkbox import Checkbox
+from .input_field import InputField
+from .slider import Slider
+from .toggle import Toggle
